split icon logic from extension messaging to keep icon bloat out of freighter-api
